storage-service-py/requirements.txt
fastapi
uvicorn[standard]
motor # Async MongoDB driver
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography] # For validating JWTs
passlib[bcrypt] # For JWT validation or if any local auth elements

# File Storage (S3)
boto3
# For local file storage testing or alternative if S3 is not primary:
# aiofiles # For async local file operations

# Image Processing
Pillow # PIL fork

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-motor
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-boto3sqs # If using SQS with Boto3, check for general boto3 instrumentor
# opentelemetry-instrumentation-botocore (More general for boto3)
httpx

# Metrics
prometheus_client

# For multipart form data (file uploads)
python-multipart

# Scheduler for CleanupService (if not using Celery for this service)
apscheduler

# Other utilities
email-validator # If handling emails in any models/schemas
python-magic # For MIME type detection from file content (optional, but good)
